Move closure-code to separate file and NSArray, NSDictionary code to separate file
