+++ Sources/Decodable.swift
@@ -44,33 +44,6 @@ extension Decodable {
     }
 }
 
-extension NSDictionary {
-    public static func decode(_ j: AnyObject) throws -> NSDictionary {
-        guard let result = j as? NSDictionary else {
-            throw TypeMismatchError(expectedType: self, receivedType: j.dynamicType, object: j)
-        }
-        return result
-    }
-}
-
-extension NSArray {
-    public static func decode(_ json: AnyObject) throws -> NSArray {
-        
-        guard let result = json as? NSArray else {
-            throw TypeMismatchError(expectedType: self, receivedType: json.dynamicType, object: json)
-        }
-        return result
-    }
-    
-    public static func decode<T>(_ context: DecodingContext<T>) throws -> NSArray {
-        
-        guard let result = context.json as? NSArray else {
-            throw TypeMismatchError(expectedType: self, receivedType: context.json.dynamicType, object: context.json)
-        }
-        return result
-    }
-}
-
 /*
 extension Dictionary where Key: Decodable, Value: Decodable {
     public static func decode(_ j: AnyObject) throws -> Dictionary {
@@ -117,34 +90,5 @@ extension Array where Element: Decodable {
 	}
  }
  */
-/// Designed to be used with parse(json, path, decodeClosure) as the decodeClosure. Thats why it's curried and a "top-level" function instead of a function in an array extension. For everyday use, prefer using [T].decode(json) instead.
-public func decodeArray<T>(_ elementDecodeClosure: (AnyObject) throws -> T) -> (AnyObject) throws -> [T] {
-    return { json in
-        return try NSArray.decode(json).map {
-            return try elementDecodeClosure($0)
-        }
-    }
-}
 
-
-public func decodeArray<T, Parameters>(_ elementDecodeClosure: (DecodingContext<Parameters>) throws -> T) -> (DecodingContext<Parameters>) throws -> [T] {
-    return { context in
-        return try NSArray.decode(context).map {
-            var c = context
-            c.json = $0
-            return try elementDecodeClosure(c)
-        }
-    }
-}
-
-/// Designed to be used with parse(json, path, decodeClosure) as the decodeClosure. Thats why it's curried. For everyday use, prefer using [K: V].decode(json) instead (declared in Decodable.swift).
-public func decodeDictionary<K,V>(_ keyDecodeClosure: (AnyObject) throws -> K, elementDecodeClosure: (AnyObject) throws -> V) -> (json: AnyObject) throws -> [K: V] {
-    return { json in
-        var dict = [K: V]()
-        for (key, value) in try NSDictionary.decode(json) {
-            try dict[keyDecodeClosure(key)] = elementDecodeClosure(value)
-        }
-        return dict
-    }
-}
